Small bug in FreeAnchor head (#944)
* reduce 1e-12 to make difference between variables

If 1e-12 is used to make a difference between t1 and t2, then later (t2-t1) might become 0 as 1e-12 can't affect to the number (at least true in pytorch, verified in v1.5.0+).

* Revert "reduce 1e-12 to make difference between variables"

This reverts commit 33ec7215d22be59bf0012dea748f5cacc33a0fb8.

* reduce 1e-12 to make difference between variables

If 1e-12 is used to make a difference between t1 and t2, then later (t2-t1) might become 0 as 1e-12 can't affect to the number (at least true in pytorch, verified in v1.5.0+). Not sure whether 1e-7 or 1e-8 worked. But 1e-6 did.